Move CommandLineParser.java to module jolie-cli
