+++ extensions/jolie-embedding-legacy/pom.xml
@@ -67,5 +67,10 @@
             <artifactId>jolie</artifactId>
             <version>${jolie.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>jolie-cli</artifactId>
+            <version>${jolie.version}</version>
+        </dependency>
     </dependencies>
 </project>
