+++ extensions/jolie-embedding-legacy/src/main/java/jolie/embedding/jolie/JolieServiceLoader.java
@@ -21,8 +21,8 @@
 
 package jolie.embedding.jolie;
 
-import jolie.CommandLineException;
-import jolie.CommandLineParser;
+import jolie.cli.CommandLineException;
+import jolie.cli.CommandLineParser;
 import jolie.Interpreter;
 import jolie.runtime.Value;
 import jolie.runtime.embedding.EmbeddedServiceLoader;
