+++ javaServices/coreJavaServices/pom.xml
@@ -470,6 +470,11 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>jolie-cli</artifactId>
+            <version>${jolie.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>jolie-js</artifactId>
             <version>${jolie.version}</version>
         </dependency>
