+++ jolie-cli/src/main/java/jolie/cli/CommandLineException.java
@@ -19,7 +19,7 @@
  *   For details about the authors of this software, see the AUTHORS file. *
  ***************************************************************************/
 
-package jolie;
+package jolie.cli;
 
 import jolie.lang.Constants;
