+++ launchers/unix/joliec
@@ -1,3 +1,3 @@
 #!/bin/sh
-java -cp $JOLIE_HOME/jolie.jar:$JOLIE_HOME/lib/libjolie.jar:$JOLIE_HOME/lib/automaton.jar:$JOLIE_HOME/tools/joliec.jar jolie.compiler.Main -l ./lib/*:$JOLIE_HOME/lib:$JOLIE_HOME/javaServices/*:$JOLIE_HOME/extensions/* -i $JOLIE_HOME/include -p $JOLIE_HOME/packages "$@"
+java -cp $JOLIE_HOME/jolie-cli.jar:$JOLIE_HOME/jolie.jar:$JOLIE_HOME/lib/libjolie.jar:$JOLIE_HOME/lib/automaton.jar:$JOLIE_HOME/tools/joliec.jar jolie.compiler.Main -l ./lib/*:$JOLIE_HOME/lib:$JOLIE_HOME/javaServices/*:$JOLIE_HOME/extensions/* -i $JOLIE_HOME/include -p $JOLIE_HOME/packages "$@"
