+++ support/metaservice-java/src/main/java/joliex/metaservice/impl/EmbeddedMetaService.java
@@ -26,8 +26,8 @@ import java.io.IOException;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
-import jolie.CommandLineException;
-import jolie.CommandLineParser;
+import jolie.cli.CommandLineException;
+import jolie.cli.CommandLineParser;
 import jolie.Interpreter;
 import jolie.net.CommChannel;
 import joliex.metaservice.MetaService;
