+++ tools/jolie2java/pom.xml
@@ -58,6 +58,11 @@
 			<version>${jolie.version}</version>
 		</dependency>
 		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>jolie-cli</artifactId>
+			<version>${jolie.version}</version>
+		</dependency>
+		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<version>4.13.1</version>
