+++ tools/jolie2plasma/src/main/java/joliex/plasma/Jolie2Plasma.java
@@ -21,8 +21,8 @@
 
 package joliex.plasma;
 
-import jolie.CommandLineException;
-import jolie.CommandLineParser;
+import jolie.cli.CommandLineException;
+import jolie.cli.CommandLineParser;
 import jolie.JolieURLStreamHandlerFactory;
 import jolie.lang.CodeCheckingException;
 import jolie.lang.parse.ParserException;
