+++ tools/jolie2surface/src/main/java/joliex/surface/JolieToSurfaceCommandLineParser.java
@@ -21,8 +21,8 @@ package joliex.surface;
 
 import java.io.IOException;
 import java.util.List;
-import jolie.CommandLineException;
-import jolie.CommandLineParser;
+import jolie.cli.CommandLineException;
+import jolie.cli.CommandLineParser;
 
 /**
  *
