+++ tools/jolie2wsdl/pom.xml
@@ -56,7 +56,12 @@
 			<artifactId>jolie</artifactId>
 			<version>${jolie.version}</version>
 		</dependency>
-				<dependency>
+		<dependency>
+			<groupId>${project.groupId}</groupId>
+			<artifactId>jolie-cli</artifactId>
+			<version>${jolie.version}</version>
+		</dependency>
+		<dependency>
 			<groupId>${project.groupId}</groupId>
 			<artifactId>jolie-xml</artifactId>
 			<version>${jolie.version}</version>
