+++ tools/jolie2wsdl/src/main/java/joliex/wsdl/Jolie2WsdlCommandLineParser.java
@@ -2,8 +2,8 @@ package joliex.wsdl;
 
 import java.io.IOException;
 import java.util.List;
-import jolie.CommandLineException;
-import jolie.CommandLineParser;
+import jolie.cli.CommandLineException;
+import jolie.cli.CommandLineParser;
 
 public class Jolie2WsdlCommandLineParser extends CommandLineParser {
